gestion d'ereure

+++ proc.py
@@ -30,24 +30,50 @@ def get_process():#{
 def kill_process():#{
     dirs = os.listdir("/proc")
     for pid in dirs:#{
-        pidDel = input("entrer le pid du processus a kill")
-        name = get_pname(pidDel)
-        p = psutil.Process(pidDel)
-        p.terminate()
-        print "vous avez kill {}".format(name)
-        break
+    	try:#{
+        	pidDel = input("PID du processus à Kill : ")
+        	if psutil.pid_exists(pidDel):#{
+        		try:#{
+        			name = get_pname(pidDel)
+        			p = psutil.Process(pidDel)
+        			p.terminate()
+        			print "Vous avez kill le processus :", name.rstrip('\n')
+        			break
+                #}
+        		except psutil.AccessDenied:
+        			print "Vous n'avez pas les droits pour kill le processus", name.rstrip('\n')
+            #}
+        	else:
+        		print "Aucun processus avec le pid \"{}\" en cours.".format(pidDel)
+        except (NameError, TypeError):
+        	print "Vous n'avez pas entrer un nombre entier."
+        #}
     #}
     return (0)
 #}
 
 def statuts_process():#{
-    pidStat = input("entrer le pid du processus a voir en details")
-    pStat = psutil.Process(pidStat)
-    parentId = psutil.Process(pidStat).ppid()
-    print (parentId)
-    print (pidStat)
-    print get_pname(pidStat)
-    print "cpu usage", (pStat.cpu_percent(interval=1))
-    print "memories usage", round((pStat.memory_percent()), 1)
-    return (0)
+	try:#{
+		pidStat = input("PID du processus pour voir plus de détails : ")
+		if psutil.pid_exists(pidStat):#{
+  			pStat = psutil.Process(pidStat)
+  			parentId = psutil.Process(pidStat).ppid()
+  			print "PPID :", (parentId)
+  			print "PID :", (pidStat)
+  			print "Username :", pStat.username()
+  			try:#{
+  				print pStat.cwd()
+  				print "Cmd name :", get_pname(pidStat).rstrip('\n')
+  				print "Statut :", pStat.status()
+  				print "CPU usage :", (pStat.cpu_percent(interval=1)), "%"
+  				print "Memories usage :", round((pStat.memory_percent()), 1), "%"
+            #}
+  		    	except psutil.AccessDenied:
+  		    		print "Vous n'avez pas les droits sur ce processus pour voir plus de détails"
+  		else:
+  			print "Aucun processus avec le PID \"{}\" en cours.".format(pidStat)
+	#}
+    except (NameError, TypeError):
+		print "Vous n'avez pas entrer un nombre entier."
+  	return (0)
 #}
